dashboard/requirements.txt
streamlit
pandas
google-cloud-bigquery
db-dtypes
pyarrow
requests
